tests: staff controllers
