--- tests/Todo/Feature/Http/Controllers/Staff/ArticleControllerTest.php
@@ -1,125 +1,0 @@
-<?php
-
-namespace Tests\Feature\Http\Controllers\Staff;
-
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
-
-/**
- * @see \App\Http\Controllers\Staff\ArticleController
- */
-class ArticleControllerTest extends TestCase
-{
-    use RefreshDatabase;
-
-    /**
-     * @test
-     */
-    public function create_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.articles.create'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.article.create');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function destroy_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $article = factory(\App\Models\Article::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->delete(route('staff.articles.destroy', ['id' => $article->id]));
-
-        $response->assertRedirect(withSuccess('Article has successfully been deleted'));
-        $this->assertDeleted($staff);
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function edit_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $article = factory(\App\Models\Article::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.articles.edit', ['id' => $article->id]));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.article.edit');
-        $response->assertViewHas('article');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function index_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.articles.index'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.article.index');
-        $response->assertViewHas('articles');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function store_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->post(route('staff.articles.store'), [
-            // TODO: send request data
-        ]);
-
-        $response->assertRedirect(withErrors($v->errors()));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function update_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $article = factory(\App\Models\Article::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->post(route('staff.articles.update', ['id' => $article->id]), [
-            // TODO: send request data
-        ]);
-
-        $response->assertRedirect(withErrors($v->errors()));
-
-        // TODO: perform additional assertions
-    }
-
-    // test cases...
-}
