--- tests/Todo/Feature/Http/Controllers/Staff/CategoryControllerTest.php
@@ -1,125 +1,0 @@
-<?php
-
-namespace Tests\Feature\Http\Controllers\Staff;
-
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
-
-/**
- * @see \App\Http\Controllers\Staff\CategoryController
- */
-class CategoryControllerTest extends TestCase
-{
-    use RefreshDatabase;
-
-    /**
-     * @test
-     */
-    public function create_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.categories.create'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.category.create');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function destroy_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $category = factory(\App\Models\Category::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->delete(route('staff.categories.destroy', ['id' => $category->id]));
-
-        $response->assertRedirect(withSuccess('Category Successfully Deleted'));
-        $this->assertDeleted($staff);
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function edit_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $category = factory(\App\Models\Category::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.categories.edit', ['id' => $category->id]));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.category.edit');
-        $response->assertViewHas('category');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function index_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.categories.index'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.category.index');
-        $response->assertViewHas('categories');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function store_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->post(route('staff.categories.store'), [
-            // TODO: send request data
-        ]);
-
-        $response->assertRedirect(withErrors($v->errors()));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function update_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $category = factory(\App\Models\Category::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->patch(route('staff.categories.update', ['id' => $category->id]), [
-            // TODO: send request data
-        ]);
-
-        $response->assertRedirect(withErrors($v->errors()));
-
-        // TODO: perform additional assertions
-    }
-
-    // test cases...
-}
